Addressing maintainability issues.
