+++ api/src/main/java/ca/bc/gov/educ/api/trax/service/AuthorityContactCreatedService.java
@@ -1,22 +1,13 @@
 package ca.bc.gov.educ.api.trax.service;
 
 import ca.bc.gov.educ.api.trax.constant.EventType;
-import ca.bc.gov.educ.api.trax.constant.FieldType;
 import ca.bc.gov.educ.api.trax.model.dto.AuthorityContact;
-import ca.bc.gov.educ.api.trax.model.dto.GradStatusEventPayloadDTO;
-import ca.bc.gov.educ.api.trax.model.dto.SchoolContact;
 import ca.bc.gov.educ.api.trax.model.entity.Event;
-import ca.bc.gov.educ.api.trax.model.entity.TraxStudentEntity;
-import jakarta.persistence.EntityManager;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.lang3.tuple.Pair;
 import org.springframework.stereotype.Service;
-
-import java.util.Map;
-
 @Service
 @Slf4j
-public class AuthorityContactCreatedService extends EventCommonService<AuthorityContact> {
+public class AuthorityContactCreatedService extends EventBaseService<AuthorityContact> {
 
     @Override
     public void processEvent(final AuthorityContact districtContact, Event event) {
@@ -30,13 +21,4 @@ public class AuthorityContactCreatedService extends EventCommonService<Authority
         return EventType.CREATE_AUTHORITY_CONTACT.toString();
     }
 
-    @Override
-    public void specialHandlingOnUpdateFieldsMap(Map<String, Pair<FieldType, Object>> updateFieldsMap, TraxStudentEntity traxStudentEntity, GradStatusEventPayloadDTO gradStatusUpdate) {
-
-    }
-
-    @Override
-    public EntityManager getEntityManager() {
-        return null;
-    }
 }
