+++ api/src/main/java/ca/bc/gov/educ/api/trax/service/AuthorityContactUpdatedService.java
@@ -1,21 +1,13 @@
 package ca.bc.gov.educ.api.trax.service;
 
 import ca.bc.gov.educ.api.trax.constant.EventType;
-import ca.bc.gov.educ.api.trax.constant.FieldType;
 import ca.bc.gov.educ.api.trax.model.dto.AuthorityContact;
-import ca.bc.gov.educ.api.trax.model.dto.GradStatusEventPayloadDTO;
 import ca.bc.gov.educ.api.trax.model.entity.Event;
-import ca.bc.gov.educ.api.trax.model.entity.TraxStudentEntity;
-import jakarta.persistence.EntityManager;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.lang3.tuple.Pair;
 import org.springframework.stereotype.Service;
-
-import java.util.Map;
-
 @Service
 @Slf4j
-public class AuthorityContactUpdatedService extends EventCommonService<AuthorityContact> {
+public class AuthorityContactUpdatedService extends EventBaseService<AuthorityContact> {
 
     @Override
     public void processEvent(final AuthorityContact districtContact, Event event) {
@@ -29,13 +21,4 @@ public class AuthorityContactUpdatedService extends EventCommonService<Authority
         return EventType.UPDATE_AUTHORITY_CONTACT.toString();
     }
 
-    @Override
-    public void specialHandlingOnUpdateFieldsMap(Map<String, Pair<FieldType, Object>> updateFieldsMap, TraxStudentEntity traxStudentEntity, GradStatusEventPayloadDTO gradStatusUpdate) {
-
-    }
-
-    @Override
-    public EntityManager getEntityManager() {
-        return null;
-    }
 }
