+++ api/src/main/java/ca/bc/gov/educ/api/trax/service/CodeService.java
@@ -8,8 +8,8 @@ import ca.bc.gov.educ.api.trax.model.transformer.GradCountryTransformer;
 import ca.bc.gov.educ.api.trax.model.transformer.GradProvinceTransformer;
 import ca.bc.gov.educ.api.trax.repository.GradCountryRepository;
 import ca.bc.gov.educ.api.trax.repository.GradProvinceRepository;
-import ca.bc.gov.educ.api.trax.util.EducGradTraxApiConstants;
 import ca.bc.gov.educ.api.trax.util.GradValidation;
+import jakarta.transaction.Transactional;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -17,8 +17,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.web.reactive.function.client.WebClient;
 
-import jakarta.transaction.Transactional;
-import java.util.*;
+import java.util.List;
+import java.util.Optional;
 
 @Service
 public class CodeService {
