+++ api/src/main/java/ca/bc/gov/educ/api/trax/service/SchoolContactDeletedService.java
@@ -1,21 +1,14 @@
 package ca.bc.gov.educ.api.trax.service;
 
 import ca.bc.gov.educ.api.trax.constant.EventType;
-import ca.bc.gov.educ.api.trax.constant.FieldType;
-import ca.bc.gov.educ.api.trax.model.dto.GradStatusEventPayloadDTO;
 import ca.bc.gov.educ.api.trax.model.dto.SchoolContact;
 import ca.bc.gov.educ.api.trax.model.entity.Event;
-import ca.bc.gov.educ.api.trax.model.entity.TraxStudentEntity;
-import jakarta.persistence.EntityManager;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.lang3.tuple.Pair;
 import org.springframework.stereotype.Service;
-
-import java.util.Map;
 
 @Service
 @Slf4j
-public class SchoolContactDeletedService extends EventCommonService<SchoolContact> {
+public class SchoolContactDeletedService extends EventBaseService<SchoolContact> {
 
     @Override
     public void processEvent(final SchoolContact districtContact, Event event) {
@@ -29,13 +22,4 @@ public class SchoolContactDeletedService extends EventCommonService<SchoolContac
         return EventType.DELETE_SCHOOL_CONTACT.toString();
     }
 
-    @Override
-    public void specialHandlingOnUpdateFieldsMap(Map<String, Pair<FieldType, Object>> updateFieldsMap, TraxStudentEntity traxStudentEntity, GradStatusEventPayloadDTO gradStatusUpdate) {
-
-    }
-
-    @Override
-    public EntityManager getEntityManager() {
-        return null;
-    }
 }
