+++ api/src/test/java/ca/bc/gov/educ/api/trax/controller/PsiControllerTest.java
@@ -14,7 +14,6 @@ import org.mockito.junit.MockitoJUnitRunner;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 
 @RunWith(MockitoJUnitRunner.class)
