+++ api/src/test/java/ca/bc/gov/educ/api/trax/service/AuthorityContactUpdatedServiceTest.java
@@ -19,15 +19,10 @@ public class AuthorityContactUpdatedServiceTest extends BaseReplicationServiceTe
         this.authorityContactUpdatedService.processEvent(request, event);
         var result = this.replicationTestUtils.getEventRepository().findById(event.getReplicationEventId());
         if(result.isPresent()){
-            Assert.assertEquals(result.get().getEventStatus(), "PROCESSED");
+            Assert.assertEquals("PROCESSED", result.get().getEventStatus());
         } else {
             fail("UPDATE_AUTHORITY_CONTACT failed to process");
         }
-    }
-
-    @Test
-    public void testGetEntityManager_expectNull() {
-        Assert.assertNull(this.authorityContactUpdatedService.getEntityManager());
     }
 
 }
