+++ api/src/test/java/ca/bc/gov/educ/api/trax/service/SchoolContactCreatedServiceTest.java
@@ -19,15 +19,10 @@ public class SchoolContactCreatedServiceTest extends BaseReplicationServiceTest 
         this.schoolContactCreatedService.processEvent(request, event);
         var result = this.replicationTestUtils.getEventRepository().findById(event.getReplicationEventId());
         if(result.isPresent()){
-            Assert.assertEquals(result.get().getEventStatus(), "PROCESSED");
+            Assert.assertEquals("PROCESSED", result.get().getEventStatus());
         } else {
             fail("CREATE_SCHOOL_CONTACT failed to process");
         }
-    }
-
-    @Test
-    public void testGetEntityManager_expectNull() {
-        Assert.assertNull(this.schoolContactCreatedService.getEntityManager());
     }
 
 }
