--- api/src/test/java/ca/bc/gov/educ/api/trax/support/MockConfiguration.java
@@ -1,19 +1,0 @@
-package ca.bc.gov.educ.api.trax.support;
-
-import ca.bc.gov.educ.api.trax.messaging.NatsConnection;
-import ca.bc.gov.educ.api.trax.messaging.jetstream.Publisher;
-import ca.bc.gov.educ.api.trax.messaging.jetstream.Subscriber;
-import io.nats.client.Connection;
-import org.mockito.Mockito;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Primary;
-import org.springframework.context.annotation.Profile;
-
-@Profile("test")
-@Configuration
-public class MockConfiguration {
-
-
-
-}
